// apps/web-svelte/src/lib/api/generated/model/videoPlaybackSource.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { VideoPlaybackMethod } from './videoPlaybackMethod';
import type { VideoPlaybackStream } from './videoPlaybackStream';
import type { VideoTranscodingInfo } from './videoTranscodingInfo';

export interface VideoPlaybackSource {
  id: string;
  /** @nullable */
  container: string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
     */
  durationSeconds: number | string | null;
  method: VideoPlaybackMethod;
  url: string;
  supportsTranscoding: boolean;
  streams: VideoPlaybackStream[];
  transcoding: null | VideoTranscodingInfo;
}