// apps/web-svelte/src/lib/api/generated/model/videoPlaybackStream.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface VideoPlaybackStream {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  index: number | string;
  type: string;
  /** @nullable */
  codec: string | null;
  /** @nullable */
  language: string | null;
  /** @nullable */
  displayTitle: string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  width: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  height: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
     */
  averageFrameRate: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  bitRate: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  sampleRate: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  channels: number | string | null;
  isDefault?: boolean;
  isForced?: boolean;
  /** @nullable */
  videoRange?: string | null;
  /** @nullable */
  videoRangeType?: string | null;
  /** @nullable */
  pixelFormat?: string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  bitDepth?: number | string | null;
  /** @nullable */
  colorRange?: string | null;
  /** @nullable */
  colorSpace?: string | null;
  /** @nullable */
  colorTransfer?: string | null;
  /** @nullable */
  colorPrimaries?: string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  dvProfile?: number | string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  dvLevel?: number | string | null;
  /** @nullable */
  rpuPresentFlag?: boolean | null;
  /** @nullable */
  elPresentFlag?: boolean | null;
  /** @nullable */
  blPresentFlag?: boolean | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  dvBlSignalCompatibilityId?: number | string | null;
  hdr10PlusPresentFlag?: boolean;
}